Add controllers Factory
This is the first step towards a more flexible controllers/reconcilers
initialization. Moving the construction out of main and integrating with
the dependencies factory.

The next step would be to remove the provider specific reconciler
construction out the main cluster reconciliation loop.

Removed the vsphere machine config controller since it wasn't being
used.